+++ mitra_activitypub/src/builders/collection.rs
@@ -20,8 +20,9 @@ pub struct OrderedCollection {
     #[serde(skip_serializing_if = "Option::is_none")]
     attributed_to: Option<String>,
 
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    ordered_items: Vec<JsonValue>,
+    // Can be serialized into an empty array
+    #[serde(skip_serializing_if = "Option::is_none")]
+    ordered_items: Option<Vec<JsonValue>>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     first: Option<String>,
@@ -43,7 +44,7 @@ impl OrderedCollection {
             id: collection_id,
             object_type: ORDERED_COLLECTION.to_string(),
             attributed_to: None,
-            ordered_items: vec![],
+            ordered_items: None,
             first: first_page_id,
             total_items,
         }
@@ -58,7 +59,7 @@ impl OrderedCollection {
             id: collection_id,
             object_type: ORDERED_COLLECTION.to_string(),
             attributed_to: None,
-            ordered_items: items,
+            ordered_items: Some(items),
             first: None,
             total_items: None,
         }
@@ -73,7 +74,7 @@ impl OrderedCollection {
             id: collection_page_id,
             object_type: ORDERED_COLLECTION_PAGE.to_string(),
             attributed_to: None,
-            ordered_items: items,
+            ordered_items: Some(items),
             first: None,
             total_items: None,
         }
@@ -82,5 +83,70 @@ impl OrderedCollection {
     pub fn with_attributed_to(mut self, attributed_to: &str) -> Self {
         self.attributed_to = Some(attributed_to.to_owned());
         self
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use serde_json::{json, to_value};
+    use super::*;
+
+    #[test]
+    fn test_build_collection_root() {
+        let collection_id = "https://example.social/collection";
+        let collection = OrderedCollection::new(
+            collection_id.to_owned(),
+            None,
+            Some(20),
+        );
+        let expected_value = json!({
+            "@context": [
+                "https://www.w3.org/ns/activitystreams",
+                "https://w3id.org/security/v1",
+                "https://w3id.org/security/data-integrity/v2",
+                {
+                    "Hashtag": "as:Hashtag",
+                    "sensitive": "as:sensitive",
+                    "toot": "http://joinmastodon.org/ns#",
+                    "Emoji": "toot:Emoji"
+                },
+            ],
+            "type": "OrderedCollection",
+            "id": "https://example.social/collection",
+            "totalItems": 20
+        });
+        assert_eq!(
+            to_value(collection).unwrap(),
+            expected_value,
+        );
+    }
+
+    #[test]
+    fn test_build_collection_with_no_items() {
+        let collection_id = "https://example.social/collection";
+        let collection = OrderedCollection::new_with_items(
+            collection_id.to_owned(),
+            vec![],
+        );
+        let expected_value = json!({
+            "@context": [
+                "https://www.w3.org/ns/activitystreams",
+                "https://w3id.org/security/v1",
+                "https://w3id.org/security/data-integrity/v2",
+                {
+                    "Hashtag": "as:Hashtag",
+                    "sensitive": "as:sensitive",
+                    "toot": "http://joinmastodon.org/ns#",
+                    "Emoji": "toot:Emoji"
+                },
+            ],
+            "type": "OrderedCollection",
+            "id": "https://example.social/collection",
+            "orderedItems": []
+        });
+        assert_eq!(
+            to_value(collection).unwrap(),
+            expected_value,
+        );
     }
 }